Toggle and Hold Button Functionality for Intake Motor
An intake motor group was added that is used to control a flap on the robot that discards incorrectly retrieved blocks. The driver can control this motor in two methods; pressing button R1 once on the controller or holding button R2. After a single press of R1, the flap will open and remain open until a subsequent press. If holding the button R2, the flap will stay open as long as R2 is held.